extract terrain controls and gui state to separate files
