+++ gui-imgui/CMakeLists.txt
@@ -37,6 +37,7 @@ target_link_libraries(imgui PUBLIC glfw)
 
 add_executable(gui-imgui
     src/main.cpp
+    src/terrainControls.cpp
 )
 
 target_include_directories(gui-imgui PUBLIC
