+++ gui-imgui/src/terrainControls.h
@@ -1,0 +1,7 @@
+#pragma once
+#include "guiState.h"
+
+class TerrainControlsWindow {
+public:
+    void render(GuiState& state);
+};
